requirements.txt
tensorflow==2.15.0
streamlit==1.29.0
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
Pillow==10.1.0
opencv-python==4.8.1.78
tf-keras-vis==0.8.7
tqdm==4.66.1
ipykernel
keras-tuner
mlflow==2.8.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6  